cos_project3/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
cycler==0.10.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-filter==21.1
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.0.0
dnspython==2.1.0
eventlet==0.32.0
fake-useragent==0.1.11
greenlet==1.1.1
gunicorn @ file:///tmp/build/80754af9/gunicorn_1617120647592/work
idna==3.2
importlib-metadata==4.8.1
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
lxml==4.6.3
matplotlib==3.3.4
mysqlclient @ file:///tmp/build/80754af9/mysqlclient_1609786257049/work
numpy==1.19.5
opencv-contrib-python==4.5.3.56
pandas==1.1.5
parse==1.19.0
Pillow==8.3.2
prompt-toolkit==3.0.20
pyee==8.2.2
PyJWT==2.1.0
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
pytesseract==0.3.7
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
requests==2.26.0
requests-html==0.10.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.2
threadpoolctl==2.2.0
tqdm==4.62.3
urllib3==1.26.7
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
zipp==3.5.0
